Move implementation of the "generate" command out of Program and into a separate class
