+++ src/ChangeLog/Commands/ICommand.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Grynwald.ChangeLog.Commands
+{
+    internal interface ICommand
+    {
+        Task<int> RunAsync();
+    }
+}
